requirements.txt
flask
google-cloud-storage
google-api-python-client==1.6.2
scikit-learn==0.19.1
pandas==0.22.0
scipy==1.0.0
numpy==1.17

